% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim_discr_raw.R
\name{anim_discr_raw}
\alias{anim_discr_raw}
\title{Function to create a raw animation for discrete data}
\usage{
anim_discr_raw(
  transition_df,
  path_file_name,
  anim_width = 500,
  anim_height = 500,
  anim_res = 500,
  column = "v_plot1",
  plot_limits,
  title = "",
  title_size = 18,
  title_face = "bold.italic",
  fill_colors = NA,
  color_pal = rainbow,
  edge_color = NA,
  edge_width = 0.5,
  na_values = FALSE,
  na_color = "grey",
  na_label = "No Data",
  legend_rows = 1,
  legend_spacing = 0.4,
  legend_order = names(fill_colors),
  legend_text = 15,
  adding_outline = FALSE,
  adding_outline_borders = FALSE,
  outline_gdf = NA,
  outline_calculate_union = TRUE,
  outline_color = "darkgrey",
  outline_width = 0.5,
  outline_borders_color = "darkgrey",
  outline_borders_width = 0.5
)
}
\arguments{
\item{transition_df}{A transition data.frame (result of create_transition function)}

\item{path_file_name}{The path and file name (.gif) of the raw animation that gets created (character)}

\item{anim_width}{The width of the animation frames (in px) (numeric)}

\item{anim_height}{The height of the animation frames (in px) (numeric)}

\item{anim_res}{The nominal resolution in ppi (numeric)}

\item{column}{The name of the column in transition_df that contains the discrete data to be plotted (character)}

\item{plot_limits}{the limits of the plot (list of two vectors containing two values), usually result of define_limits()}

\item{title}{The title of the plot (character)}

\item{title_size}{The size of the title (numeric)}

\item{title_face}{The typographical emphasis of the title (character)}

\item{fill_colors}{The colors to fill the polygons with (named character vector)
(example: c("A"="beige", "B"="#8d0613", "C"="darkorange"))}

\item{color_pal}{The color palette to use for fill_colors (function). Only necessary if no fill_colors are specified!}

\item{edge_color}{The color of the polygon edges (character)}

\item{edge_width}{The width of the polygon edges (numeric)}

\item{na_values}{NA values present in column? (logical)}

\item{na_color}{The color to fill the NA values with (character)}

\item{na_label}{The label for NA values (character)}

\item{legend_rows}{The number of rows in the legend (numeric)}

\item{legend_spacing}{The spacing between legend items (numeric)}

\item{legend_order}{The order of the legend items (character vector)}

\item{legend_text}{The size of the legend text (numeric)}

\item{adding_outline}{If TRUE, outline of unionized polygons is added (logical)}

\item{adding_outline_borders}{If TRUE, outlines of every polygon/feature are added (logical)}

\item{outline_gdf}{A sf object that either contains single polygons of the gdf or union of polygons}

\item{outline_calculate_union}{Only relevant if adding_outline = TRUE: If outline_calculate_union = FALSE, outline_gdf already contains the unionized polygons (logical).
Therefore, if TRUE when creating the animation, it will be calculated anew for each frame!}

\item{outline_color}{If adding_outline = TRUE: The color of the unionized outline (character)}

\item{outline_width}{If adding_outline = TRUE: The width of the unionized outline (numeric)}

\item{outline_borders_color}{If adding_outline_borders = TRUE: The color of the polygon/feature outlines (character)}

\item{outline_borders_width}{If adding_outline_borders = TRUE: The width of the polygon/feature outlines (numeric)}
}
\value{
None, but saves a .gif file at the specified path_file_name
}
\description{
\code{anim_discr_raw} plots a ggplot of a all intermediary states of a transition df,
fills the polygons according to a discrete data column and creates a raw animation
that gets saved.
}
\author{
Georg Starz, Anna Bischof
}
