% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim_finalize.R
\name{anim_finalize}
\alias{anim_finalize}
\title{Function to fine-tune a raw animation}
\usage{
anim_finalize(
  anim_raw,
  anim_path_file,
  fps_anim = 10,
  delay_anim = TRUE,
  delay_frames = 60
)
}
\arguments{
\item{anim_raw}{The path and name of the raw animation file (.gif) (character)}

\item{anim_path_file}{The path and name of the final animation file (.gif) (character)}

\item{fps_anim}{The frames per second of the animation (numeric)}

\item{delay_anim}{Should the animation have a delay at the beginning and end? (logical)}

\item{delay_frames}{The number of frames of the beginning and end delay (numeric)}
}
\value{
None, but saves a .gif file at the specified path_file_name
}
\description{
\code{anim_finalize} loads a raw animation, defines the fps and enables a delay at the beginning
and end before the new animation gets saved.
}
\author{
Georg Starz
}
