% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bubbles.R
\name{create_bubbles}
\alias{create_bubbles}
\title{Create bubble geodataframe according to a column}
\usage{
create_bubbles(
  merged_gdf,
  col_name,
  geom_name = "geometry",
  circle_share = 1,
  change_parameter = 0.05
)
}
\arguments{
\item{merged_gdf}{The geodataframe which contains the column on which the size of the bubbles should depend (sf data.frame)}

\item{col_name}{The name of the column on which the size of the bubbles should depend (character)}

\item{geom_name}{The name of the geometry column in merged_gdf (default: "geometry")}

\item{circle_share}{The initial share of the total area of all bubbles created to the original area of merged_gdf (default: 1)}

\item{change_parameter}{A relative parameter that determines the change in circle_share when resizing the bubbles (default: 0.05)}
}
\value{
The bubble geodataframe
}
\description{
\code{create_bubbles} allows the user to create a bubble geodataframe in an individually customizable and iterative way.
The size of the bubbles depends on a column in the input geodataframe.
}
\author{
Georg Starz
}
