% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transition.R
\name{create_transition}
\alias{create_transition}
\title{Create transition between two geodataframes with different geometries}
\usage{
create_transition(
  gdf,
  bubble_gdf,
  color_col,
  bubble_col,
  id_col = "id",
  frame_number = 40,
  gdf_geom_name = "geometry",
  bubble_geom_name = "geometry"
)
}
\arguments{
\item{gdf}{The initial geodataframe (sf data.frame)}

\item{bubble_gdf}{The (bubble) geodataframe which should be displayed at the end of the animation}

\item{color_col}{The name of the column in gdf and bubble_gdf that is used for coloring in the subsequent animation (character)
(column must have the same name in both geodataframes!)}

\item{bubble_col}{The name of the column in bubble_gdf that was used to create the new (bubble) geometries (character)}

\item{id_col}{The name of the ID column in gdf and bubble_gdf (column must have the same name in both geodataframes!) (default: "id")}

\item{frame_number}{The number of frames in the transition (default: 40)}

\item{gdf_geom_name}{The name of the geometry column in gdf (default: "geometry")}

\item{bubble_geom_name}{The name of the geometry column in bubble_gdf (default: "geometry")}
}
\value{
A data.frame containing intermediary states of the transition
}
\description{
\code{create_transition} creates the transition (using tween_sf) for a subsequent animation between two geodataframes with different geometries, but partly the same IDs.
}
\author{
Georg Starz
}
