% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_limits.R
\name{define_limits}
\alias{define_limits}
\title{Defines plot limits of the plots / final animation}
\usage{
define_limits(
  data_start,
  data_end = data_start,
  parameter_w = 0.999,
  parameter_e = 1.001,
  parameter_s = 0.999,
  parameter_n = 1.001
)
}
\arguments{
\item{data_start}{A geodataframe (sf data.frame)}

\item{data_end}{A geodataframe (sf data.frame), if not specified data_start}

\item{parameter_w}{The parameter for the west limit (numeric). Decrease to have larger plot limits.}

\item{parameter_e}{The parameter for the east limit (numeric). Increase to have larger plot limits.}

\item{parameter_s}{The parameter for the south limit (numeric). Decrease to have larger plot limits.}

\item{parameter_n}{The parameter for the north limit (numeric). Increase to have larger plot limits.}
}
\value{
A lists containing two vectors of two values each.
}
\description{
\code{define_limits} defines the plot limits of the final animation, or any two plots that should have the same limits. Also works for a single plot.
}
\author{
Anna Bischof, Georg Starz
}
