% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sim_change.R
\name{find_sim_change}
\alias{find_sim_change}
\title{Find similar values in two dataframe columns and change them}
\usage{
find_sim_change(
  df_main,
  df_main_col,
  df_change,
  df_change_col,
  sim = "sim",
  thresh = 0.5
)
}
\arguments{
\item{df_main}{A dataframe (data.frame). This df remains unchanged}

\item{df_main_col}{The name of the column in df_main that is used to compare similar values (character)}

\item{df_change}{A dataframe (data.frame). This df will be changed.}

\item{df_change_col}{The name of the column in df_change that is used to compare similar values (character)}

\item{sim}{Can be "sim", "jc" (Jaccard Similarity), or "jw" (Jaro Winkler Similarity) (character vector)}

\item{thresh}{Threshold for similarity measure (numeric)}
}
\value{
Updated df_change dataframe.
}
\description{
\code{find_sim_change} finds similar values based on different possible similarity measures in two columns of two dataframes.
It then prompts the user to decide whether the values in one dataframe should be replaced by the values in the other dataframe.
The function can also be used multiple times with different similarity measures.
}
\author{
Anna Bischof
}
