% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_df_df.R
\name{merge_df_df}
\alias{merge_df_df}
\title{Merge two dataframes}
\usage{
merge_df_df(
  df_left,
  id_left,
  df_right,
  id_right = id_left,
  cols_to_keep = "ALL"
)
}
\arguments{
\item{df_left}{A dataframe (data.frame)}

\item{id_left}{The name of the column in df_left that will be used to merge (character)}

\item{df_right}{A dataframe (data.frame)}

\item{id_right}{The name of the column in df_right that will be used to merge (character)}

\item{cols_to_keep}{The columns that should be kept in the merged df (character vector)}
}
\value{
The merged dataframe
}
\description{
\code{merge_df_df} merges a df and df with a inner join and checks if/what data would be lost in the right df
}
\author{
Anna Bischof
}
