% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_df_df.R, R/merge_gd_df.R
\name{merge_gd_df}
\alias{merge_gd_df}
\title{Merge two dataframes}
\usage{
merge_gd_df(
  gdf_left,
  id_left,
  df_right,
  id_right = id_left,
  cols_to_keep = "ALL"
)

merge_gd_df(
  gdf_left,
  id_left,
  df_right,
  id_right = id_left,
  cols_to_keep = "ALL"
)
}
\arguments{
\item{gdf_left}{The geodataframe (sf data.frame)}

\item{id_left}{The name of the column in the gdf that will be used to merge (character)}

\item{df_right}{The dataframe (data.frame)}

\item{id_right}{The name of the column in the df that will be used to merge (character)}

\item{cols_to_keep}{The columns that should be kept in the merged gdf (character vector)}

\item{df_left}{A dataframe (data.frame)}
}
\value{
The merged dataframe

The merged geodataframe
}
\description{
\code{merge_df_df} merges a df and df with a inner join and checks if/what data would be lost in the right df

\code{merge_gd_df} merges a gdf and df with a left join and checks if/what data would be lost in the right df
}
\examples{

spain_gdp <- merge_df_df(spain, "Texto", gdp_1, "CCAA", c("PIB_Per_Capita_EURO", "PIB_anual_EURO"))


spain_gdp <- merge_gd_df(spain, "Texto", gdp_1, "CCAA", c("PIB_Per_Capita_EURO", "PIB_anual_EURO"))

}
\author{
Anna Bischof
}
