% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_identify.R
\name{outlier_identify}
\alias{outlier_identify}
\title{Identify/delete spatial outliers in  vector geodata}
\usage{
outlier_identify(
  geodata,
  id_col = "id",
  search_parameter = 0.2,
  geom_name = "geometry",
  extra_cols = names(geodata)[!names(geodata) \%in\% c(id_col, "geometry")]
)
}
\arguments{
\item{geodata}{The geodataframe (sf data.frame)}

\item{id_col}{The name of the ID column (character) (default: "id")}

\item{search_parameter}{A relative parameter that defines the minimum distance
from the bounding box of the largest single part of geodata from which spatial
outliers are searched for by multiplying the square root of the bounding box area
with the parameter. If the parameter is increased, outliers are only searched for
further away, if it is decreased, closer (default: 0.2)}

\item{geom_name}{The name of the geometry column in geodata (default: "geometry")}

\item{extra_cols}{The columns that should be kept in the output gdf (character vector)
(default: all columns that were also in geodata)}
}
\value{
The geodataframe without the deleted / with the defined (newly created "outlier" column) spatial outliers
}
\description{
\code{outlier_identify} looks for spatial outlier features & subfeatures in vector geodata
and enables deleting these features/subfeatures or defining them as an outlier
in a newly created column (e.g., for Spain, the Canary Islands).
}
\author{
Georg Starz
}
