% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_moving.R
\name{outlier_moving}
\alias{outlier_moving}
\title{Move previously defined spatial outliers in  vector geodata}
\usage{
outlier_moving(geodata, move_parameter = 0.025)
}
\arguments{
\item{geodata}{The geodataframe (sf data.frame)}

\item{move_parameter}{A relative parameter that determines the step size of the outlier shifts.
It refers to the ratio of the size of the bounding boxes of the non-outlier part and the respective outlier.
If it is increased, the steps for shifting become larger; if it is decreased, they become smaller (default: 0.025)}
}
\value{
The geodataframe with the modified geometries for the spatial outliers
}
\description{
\code{outlier_moving} allows the user to move all features assigned with True in the
"outlier" column in an individually customizable and iterative way
}
\author{
Georg Starz
}
