% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_restructure.R
\name{outlier_restructure}
\alias{outlier_restructure}
\title{Restructure geodataframe after outlier_identify / outlier_moving}
\usage{
outlier_restructure(geodata)
}
\arguments{
\item{geodata}{The geodataframe (sf data.frame)}
}
\value{
The restructured geodataframe
}
\description{
\code{outlier_restructure} includes parts of features (name suffix "_part") that were previously
defined / moved as spatial outliers in outlier_identify / outlier_moving back into the respective feature
}
\author{
Georg Starz
}
