% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discrete_data.R
\name{plot_discr_data}
\alias{plot_discr_data}
\title{Function to plot discrete data}
\usage{
plot_discr_data(
  gdf,
  column,
  plot_limits,
  title = "",
  title_size = 18,
  title_face = "bold.italic",
  fill_colors = NA,
  color_pal = rainbow,
  edge_color = NA,
  edge_width = 0.5,
  na_values = FALSE,
  na_color = "grey",
  na_label = "No Data",
  legend_rows = 1,
  legend_spacing = 0.4,
  legend_order = names(fill_colors),
  legend_text = 15,
  adding_outline = FALSE,
  adding_outline_borders = FALSE,
  outline_gdf = NA,
  outline_calculate_union = TRUE,
  outline_color = "darkgrey",
  outline_width = 0.5,
  outline_borders_color = "darkgrey",
  outline_borders_width = 0.5
)
}
\arguments{
\item{gdf}{A geodataframe (sf data.frame)}

\item{column}{The name of the column in gdf that contains the data to be plotted (character)}

\item{plot_limits}{the limits of the plot (list of two vectors containing two values), usually result of define_limits()}

\item{title}{The title of the plot (character)}

\item{title_size}{The size of the title (numeric)}

\item{title_face}{The typographical emphasis of the title (character)}

\item{fill_colors}{The colors to fill the polygons with (named character vector)
(example: c("A"="beige", "B"="#8d0613", "C"="darkorange"))}

\item{color_pal}{The color palette to use for fill_colors (function). Only necessary if no fill_colors are specified!}

\item{edge_color}{The color of the polygon edges (character)}

\item{edge_width}{The width of the polygon edges (numeric)}

\item{na_values}{NA values present in column? (logical)}

\item{na_color}{The color to fill the NA values with (character)}

\item{na_label}{The label for NA values (character)}

\item{legend_rows}{The number of rows in the legend (numeric)}

\item{legend_spacing}{The spacing between legend items (numeric)}

\item{legend_order}{The order of the legend items (character vector)}

\item{legend_text}{The size of the legend text (numeric)}

\item{adding_outline}{If TRUE, outline of unionized polygons is added (logical)}

\item{adding_outline_borders}{If TRUE, outlines of every polygon/feature are added (logical)}

\item{outline_gdf}{A sf object that either contains single polygons of the gdf or union of polygons}

\item{outline_calculate_union}{Only relevant if adding_outline = TRUE: If outline_calculate_union = FALSE, outline_gdf already contains the unionized polygons (logical).
Therefore, if TRUE when creating the animation, it will be calculated anew for each frame!}

\item{outline_color}{If adding_outline = TRUE: The color of the unionized outline (character)}

\item{outline_width}{If adding_outline = TRUE: The width of the unionized outline (numeric)}

\item{outline_borders_color}{If adding_outline_borders = TRUE: The color of the polygon/feature outlines (character)}

\item{outline_borders_width}{If adding_outline_borders = TRUE: The width of the polygon/feature outlines (numeric)}
}
\value{
Plot of geodataframe with filled polygons according to discrete data
}
\description{
\code{plot_discr_data} plots a ggplot of a geodataframe and fills the polygons according to discrete data
}
\author{
Anna Bischof, Georg Starz
}
