% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spain_gdp.R
\docType{data}
\name{spain_gdp}
\alias{spain_gdp}
\title{GDP data of Spanish federal states}
\format{
##'spain_gdp'
A data frame with 19 rows and 4 columns:
\describe{
\item{CCAA}{Name of the state}
\item{Fecha}{Year of the data}
\item{PIB_Per_Capita_EURO}{GDP per capita in euros}
\item{PIB_anual_EURO}{Total GDP of the state in euros}
}
@source \url{https://datosmacro.expansion.com/pib/espana-comunidades-autonomas}
}
\usage{
spain_gdp
}
\description{
This dataset contains the GDP per capita and the total GDP for all Spanish federal states in 2022.
}
\keyword{datasets}
