% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spain_gpkg.R
\docType{data}
\name{spain_gpkg}
\alias{spain_gpkg}
\title{Geodata of Spanish federal states}
\format{
##'spain_gpkg'
A data frame with 19 rows and 4 columns:
\describe{
\item{Codigo}{Numeric code of the state}
\item{Texto}{Name of the state}
\item{Texto_Alt}{Name of the state in the regional language}
\item{geom}{Geometry column}
}
@source \url{https://www.arcgis.com/home/item.html?id=5f689357238847bc823a2fb164544a77#!}
}
\usage{
spain_gpkg
}
\description{
This dataset contains geodata of all Spanish federal states.
}
\keyword{datasets}
