% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gd_df.R
\name{merge_gd_df}
\alias{merge_gd_df}
\title{Merge geodataframe with dataframe}
\usage{
merge_gd_df(
  gdf_left,
  id_left,
  df_right,
  id_right = id_left,
  cols_to_keep = "ALL"
)
}
\arguments{
\item{gdf_left}{The geodataframe (sf data.frame)}

\item{id_left}{The name of the column in the gdf that will be used to merge (character)}

\item{df_right}{A dataframe (data.frame)}

\item{id_right}{The name of the column in the df that will be used to merge (character)}

\item{cols_to_keep}{The columns that should be kept in the merged gdf (character vector)}
}
\value{
The merged geodataframe
}
\description{
\code{merge_gd_df} merges a gdf and df with a left join and checks if/what data would be lost in the right df
}
\author{
Anna Bischof
}
